requirements.txt
openai
langchain
langchain-community
python-dotenv
streamlit
pdfminer.six
python-docx
chromadb
langchain-groq
